RYSK-50: Hoist hook to be used elsewhere.
