+++ packages/front-end/src/components/optionsTrading/Modals/CloseShortOptionModal/index.tsx
@@ -26,7 +26,7 @@ import {
 } from "../Shared/components/Form";
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
-import { useNotifications } from "../Shared/hooks/useNotifications";
+import { useNotifications } from "../../hooks/useNotifications";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
 import { approveAllowance, vaultSell } from "../Shared/utils/transactions";
 import { Pricing } from "./components/Pricing";
