+++ packages/front-end/src/components/optionsTrading/Modals/SellOptionModal/index.tsx
@@ -14,7 +14,7 @@ import { Disclaimer } from "../Shared/components/Disclaimer";
 import { Button, Input, Label, Wrapper } from "../Shared/components/Form";
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
-import { useNotifications } from "../Shared/hooks/useNotifications";
+import { useNotifications } from "../../hooks/useNotifications";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
 import { approveAllowance, sell } from "../Shared/utils/transactions";
 import { Filters } from "./components/Filters";
